dark mode is made
dark mode component that can be placed on navbar eventually,
please ignore a ton of import changes i had tried out a different file structure but in the end it was just more hassle so i moved it back
