+++ front-end/src/DarkMode/darkmode.css
@@ -1,0 +1,6 @@
+.darkMode{
+    display: flex;
+}
+.moon{
+    margin-top: 3px;
+}
