+++ front-end/src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
-import { UserProvider } from "./Components/UserProvider";
+import { UserProvider } from "./UserProvider";
 
 ReactDOM.render(
   <React.StrictMode>
